packages/won/pyproject.toml
[tool.poetry]
name = "won"
version = "0.0.7"
description = ""
authors = ["Mariusz Kabala <mariusz@kabala.waw.pl>"]

[tool.poetry.dependencies]
python = "^3.7"
wakeonlan = "*"
paho-mqtt = "^1.5.0"
flask = "*"
requests = "*"
python-json-logger = "^2.0.4"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[package]
include = ["*.py"]

[tool.poetry.scripts]
start = "won:start"
